세션으로 모든걸 막는 작업!

+++ jan08/jan08web/src/main/java/com/poseidon/web/Delete.java
@@ -8,6 +8,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import com.poseidon.dao.BoardDAO;
 import com.poseidon.dto.BoardDTO;
@@ -23,23 +24,29 @@ public class Delete extends HttpServlet {
 
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 //		response.getWriter().append("Served at: ").append(request.getContextPath());
-
+		HttpSession session = request.getSession();
 //  	오는 db 잡기
 //			숫자네?삭제 진행-> 보드로 이동
 //			번호 받기
 			int no = Util.str2Int(request.getParameter("no"));
+			String mid = request.getParameter("mname");
+			if(session.getAttribute("mid")==mid) {
 //			DAO 일 시키기
-			BoardDAO dao = new BoardDAO();
-			int result = dao.delete(no);
+				BoardDAO dao = new BoardDAO();
+				int result = dao.delete(no);
 //			잘 삭제되었는지 값 받기
-			System.out.println("삭제여부 : "+result);
-			
-			if(result == 1) {
+				System.out.println("삭제여부 : "+result);
+				
+				if(result == 1) {
 //				RequestDispatcher rd = request.getRequestDispatcher("delete.jsp");
 //				rd.forward(request, response);
-				response.sendRedirect("./board");
+					response.sendRedirect("./board");
+				} else {
+					response.sendRedirect("./error.jsp");
+				}
 			} else {
-				response.sendRedirect("./error.jsp");
+				System.out.println("작성자 아님");
+				response.sendRedirect("./board");
 			}
 			
 //			숫자가 아니네? => 에러 표시
